Binary input

Approved-by: Bastian Hagedorn <b_hage09@uni-muenster.de>
Approved-by: Michel Steuwer <michel.steuwer@glasgow.ac.uk>

+++ resources/example/int_example.cl
@@ -1,4 +1,4 @@
-kernel void KERNEL(global float* input, global float* output, int N) {
+kernel void KERNEL(global int* input, global int* output, int N) {
 
   int gid = get_global_id(0);
